+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Include `-webkit` prefixed versions of all flexbox properties to fix display in Safari
 - Fix image drag & drop in Chrome -- revoking blob URLs immediately now results in a broken image
 - Remove all references to `Page#path` in initialization to avoid trying to resolve the full path before the pages' parents have been assigned
+- Explicitly copy compiled assets to the new published revision using a new core publish step rather than hijack the asset compilation/resolution step. This ensures that dependencies are copied on a second publish.
 
 ## 0.2.0.beta7, released 2014-09-03